added comments/created a function to handle option array to get rid of duplicated code
